migrate common fixtures into pytest_plugins
Issues:
Fixes #845

Problem: Accessing useful fixtures in other Python
applications/libraries was too difficult.

Analysis: This commit makes generally useful fixtures
accessible to pytest via the standard pytest11 entry_point
as a plugin.

Tests: There are no new test failures in my local environment.
